requirements.txt
streamlit
langchain
pinecone-client
openai
groq-client
PyMuPDF
python-dotenv
PyPDF2
